requirements.txt
terraformpy
pyyaml
pytest-xdist
jsondiff
secrets